+++ lib/src/analyzer_plugin/analyzer_plugin_starter.dart
@@ -6,6 +6,7 @@ import 'package:analyzer_plugin/starter.dart';
 import 'analyzer_plugin.dart';
 
 void start(Iterable<String> _, SendPort sendPort) {
-  ServerPluginStarter(AnalyzerPlugin(PhysicalResourceProvider.INSTANCE))
+  ServerPluginStarter(
+          AnalyzerPlugin(resourceProvider: PhysicalResourceProvider.INSTANCE))
       .start(sendPort);
 }